pyproject.toml
[tool.poetry]
name = "cms-migration-system"
version = "0.1.0"
description = "Intelligent Content Migration System using AI to convert HTML to CMS components"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
beautifulsoup4 = "^4.12.2"
cohere = "^4.39.0"
pymongo = "^4.6.1"
python-dotenv = "^1.0.0"
lxml = "^4.9.4"
pandas = "^2.2.0"
requests = "^2.31.0"
rich = "^13.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
black = "^23.12.1"
flake8 = "^7.0.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"